+++ src/runners.js
@@ -6,20 +6,24 @@ const { firstError, containsError } = require('./util');
 const returnUndefinedOnSuccess = errors => {
   return Object.keys(errors).length === 0 ? undefined : errors;
 };
+const notEmpty = value => !isEmpty(value);
 
 /**
  *
- * @param perAttributeRules {Object} plain object describing rules for each attribute
- * @param wholeData {Object} data to be passed to validation rules as second attribute. this param is used internally,
- * you shouldn't have a need to use it.
- * @return {function(*=)}
+ * @param perAttributeRules {Object} plain object describing rules for each attribute.
+ * @param wholeData {Object} data to be passed to validation rules as second argument.
+ * this param is used internally, you shouldn't have a need to use it.
+ * @param pathToCurrentData {Array} path to be passed to validation rules as third argument.
+ * this param is used internally, you shouldn't have a need to use it.
+ * @return {function} function that you can run against your data to validate.
  */
-const createValidator = (perAttributeRules, wholeData = {}) => {
+const createValidator = (perAttributeRules, wholeData = {}, pathToCurrentData = []) => {
   return currentData => {
     wholeData = isEmpty(wholeData) ? currentData : wholeData;
     const errors = {};
     // create an array of functions that will validate each attribute
     const validators = map(perAttributeRules, (rulesForKey, keyToValidate) => {
+      const pathToCurrentKey = pathToCurrentData.concat([keyToValidate]);
       const dataToValidate = currentData[keyToValidate];
       const storeErrors = validationResult => {
         if (containsError(validationResult)) {
@@ -30,12 +34,13 @@ const createValidator = (perAttributeRules, wholeData = {}) => {
       let validateFunction;
       if (isPlainObject(rulesForKey)) {
         if (!isPlainObject(dataToValidate)) {
-          validateFunction = () => Promise.resolve(`Property ${keyToValidate} must be an object`);
+          validateFunction = () => Promise.resolve(`Property ${pathToCurrentKey.join('.')} must be an object`);
         } else {
-          validateFunction = () => createValidator(rulesForKey, wholeData)(dataToValidate).then(storeErrors);
+          validateFunction = () => createValidator(rulesForKey, wholeData, pathToCurrentKey)(dataToValidate)
+            .then(storeErrors);
         }
       } else {
-        validateFunction = () => firstError(rulesForKey)(dataToValidate, wholeData).then(storeErrors);
+        validateFunction = () => firstError(rulesForKey)(dataToValidate, wholeData, pathToCurrentKey).then(storeErrors);
       }
 
       return validateFunction;