+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 #### 0.2.3
 
 - added `each` method to `util`
+- added `path` as third argument passed to all `rules` and some `utils`